Extract Kore.Internal.NormalizedAc
